Separate HTTP healthcheck into its own .go file
